requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   google-genai
    #   httpx
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
dotenv==0.9.9
    # via dishtwin (pyproject.toml)
google-auth==2.40.3
    # via google-genai
google-genai==1.36.0
    # via dishtwin (pyproject.toml)
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via google-genai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.9
    # via google-genai
pydantic-core==2.33.2
    # via pydantic
python-dotenv==1.1.1
    # via
    #   dishtwin (pyproject.toml)
    #   dotenv
requests==2.32.5
    # via google-genai
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via anyio
tenacity==9.1.2
    # via google-genai
typing-extensions==4.15.0
    # via
    #   anyio
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via requests
websockets==15.0.1
    # via google-genai
